split stuff out into 4 files, still need to finish WallGrid class
Did make some good progess, now creates walls seperately. Plan to add
a new cell methods on the wall grid to enable drawing between centers
(if wall does not exist) and accessing the walls of an existing cell.

Then need a maze object that goes in and removes walls and makes the
grid into a proper maze.